pyproject.toml
[project]
name = "interview-analyzer"
version = "0.1.0"
description = "AI-led interview platform with resume-driven questions, scoring, and report generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "jinja2>=3.1.4",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "python-docx>=1.1.2",
  "pypdf>=5.0.1",
  "orjson>=3.10.7",
  "numpy>=1.26.4",
  "scikit-learn>=1.5.1",
  "sentence-transformers>=3.0.1",
  "FlagEmbedding>=1.3.2",
  "faster-whisper>=1.1.0",
  "rapidfuzz>=3.9.6",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
